Now generating the JSP from the database

+++ src/m2geii/forum/beans/User.java
@@ -2,7 +2,7 @@ package m2geii.forum.beans;
 
 public class User 
 {	
-	private Long id;
+	private int id;
 	
 	private String login;
 	private String pass;
@@ -11,12 +11,12 @@ public class User
 	private String secondname;
 	
 	
-	public Long getId() 
+	public int getId() 
 	{
 		return id;
 	}
 	
-	public void setId(Long id) 
+	public void setId(int id) 
 	{
 		this.id = id;
 	}
